requirements.txt
anyio==3.5.0
asgiref==3.5.0
awscli==1.22.85
botocore==1.24.30
click==8.1.1
colorama==0.4.3
docutils==0.15.2
fastapi==0.75.0
h11==0.13.0
idna==3.3
jmespath==1.0.0
pyasn1==0.4.8
pydantic==1.9.0
python-dateutil==2.8.2
python-dotenv==0.20.0
PyYAML==5.4.1
rsa==4.7.2
s3transfer==0.5.2
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
typing-extensions==4.1.1
urllib3==1.26.9
uvicorn==0.17.6
